requirements.txt
fastapi
uvicorn
langchain==0.1.16
langchain-community==0.0.38
langchain-core==0.1.53
langchain-openai==0.1.3
langchainhub
openai
boto3
python-dotenv